+++ justfile
@@ -1,10 +1,10 @@
 test:
   cargo test
-  cargo test --features=simd_json --no-default-features snapshot_tests
+  cargo test --features=simd_json --no-default-features --test image_snapshot_tests
 
 update_snapshots:
-  UPDATE_SNAPSHOTS=1 cargo test snapshot_tests
-  cargo test --features=simd_json --no-default-features snapshot_tests
+  UPDATE_SNAPSHOTS=1 cargo test --test image_snapshot_tests
+  cargo test --features=simd_json --no-default-features --test image_snapshot_tests
 
 demo:
   cargo run -p demo
